// Frontend - React/src/models/enums/component-name.enum.ts
export enum ComponentNameEnum {
    AUDIO = 'audio',
    BUTTON = 'button',
    CALENDAR = 'calendar',
    HEADING = 'heading',
    IMAGE = 'image',
    INPUT = 'input',
    LINK = 'link',
    PARAGRAPH = 'paragraph',
    PHOTO_GALLERY = 'photogallery',
    PROFILE = 'user profile',
    SLIDER = 'slider',
    VIDEO = 'video',
};